pyproject.toml
[project]
name = "pydantic-ai-project"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "gepa>=0.1.1",
    "mlflow>=3.11.1",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.77.0",
    "pydantic-ai-slim[duckduckgo]>=1.77.0",
    "python-dotenv>=1.2.2",
]
